Fix local to server deployment. Upgrade winston logger.

+++ deploy/deploy.sh
@@ -1,32 +1,41 @@
 #!/bin/sh
 
-sh deploy/build-prod-all.sh
-
-# Go down a folder, TODO To remove if we fill a /dist folder before
-cd ..
-
 baseDir=babydue-race-2
-name=$(date '+%Y%m%d%H%M%S').${baseDir}
+if [ -z "$1" ]
+  then
+    name=$(date '+%Y%m%d%H%M%S').${baseDir}
+  else
+    name=$1
+fi
+
 dirName=dist/${name}
 nameZip=$name.tar.gz
 
-echo "Creating ${baseDir}/${dirName}"
-mkdir -p ${baseDir}/${dirName}
+if [ ! -f $nameZip ]
+  then
+    sh deploy/build-prod-all.sh
 
-# Copy to dist
-echo "Copying node & /public to ${baseDir}/${dirName}"
-rsync -av --progress ${baseDir}/ ${baseDir}/${dirName}/ --exclude game --exclude data --exclude dist --exclude .git --exclude deploy/deploy.sh --exclude config --exclude cron
+    # Go down a folder
+    cd ..
+    echo "Creating ${baseDir}/${dirName}"
+    mkdir -p ${baseDir}/${dirName}
 
-mkdir -p ${baseDir}/${dirName}/game
+    # Copy to dist
+    echo "Copying node & /public to ${baseDir}/${dirName}"
+    rsync -av --progress ${baseDir}/ ${baseDir}/${dirName}/ --exclude ./game --exclude ./data --exclude ./dist --exclude .git --exclude deploy/deploy.sh --exclude config/config.js --exclude ./cron
 
-echo "Copying game to ${baseDir}/${dirName}/game"
-rsync -av --progress ${baseDir}/game/build/ ${baseDir}/${dirName}/game/
+    mkdir -p ${baseDir}/${dirName}/game
 
-cd ${baseDir}/dist
+    echo "Copying game to ${baseDir}/${dirName}/game"
+    rsync -av --progress ${baseDir}/game/build/ ${baseDir}/${dirName}/game/
 
-echo "Tarring ${name} to ${nameZip}"
-tar -zcvf "$nameZip" ${name}
+    cd ${baseDir}/dist
 
+    echo "Tarring ${name} to ${nameZip}"
+    tar -zcvf "$nameZip" ${name}
+  else
+    cd dist
+fi
 
 # Remote vars
 remoteRoot=/var/www/${baseDir}
@@ -36,25 +45,22 @@ remoteDirVersions=${remoteRoot}/www/versions
 # Setup a ssh-agent so it doesn't request SSH key password for each command
 ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "mkdir -p ${remoteDirVersions}"
 
-echo "Sending tar by ssh: sshpass -p ${BABYDUE_RACE_2_PASSWORD} scp -p ${nameZip} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST}:${remoteDirVersions}"
+echo "Sending tar by ssh: scp -p ${nameZip} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST}:${remoteDirVersions}"
 scp -i ${BABYDUE_RACE_2_SSH_KEY} -p ${nameZip} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST}:${remoteDirVersions}
 
 echo "Uncompressing"
 ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "cd ${remoteDirVersions} && tar -zxvf ${nameZip} ${name} && rm -Rf ${nameZip}"
 
-echo "Running npm install"
-ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "cd ${remoteDirVersions}/${name} && sudo npm install"
-
-# Config symlink
-echo "Setting up symlinks"
-ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "cd ${remoteDirVersions}/${name} && sh deploy/symlinks.sh ${remoteRoot}"
-
-echo "Changing owner"
-ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "chown nodejs -R ${remoteDirVersions}/${name}"
-
-# latest symlink to
-echo "cd ${remoteRoot}/www && ln -sf ${remoteDirVersions}/${name} current"
-ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "cd ${remoteRoot}/www && ln -sfn ${remoteDirVersions}/${name} current"
-
-# TODO Restart as nodejs user
-ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "cd ${remoteRoot}/www/current && pm2 restart ecosystem.config.js --env production --update-env"
+# Several tasks at once. Did in one command to overcome the SSH 6 user rate limit
+# You could check that limit by using:
+#   $ iptables -nL|grep "22\|ssh"
+ssh -i ${BABYDUE_RACE_2_SSH_KEY} ${BABYDUE_RACE_2_LOGIN}@${BABYDUE_RACE_2_HOST} "
+  echo 'Setting up symlinks'
+  cd ${remoteDirVersions}/${name} && sh deploy/symlinks.sh ${remoteRoot}
+  echo 'Changing owner'
+  chown nodejs -R ${remoteDirVersions}/${name}
+  echo 'Updating latest current symlink'
+  cd ${remoteRoot}/www && ln -sfn ${remoteDirVersions}/${name} current
+  echo 'Restarting app'
+  cd ${remoteRoot}/www/current && runuser -l nodejs -c 'pm2 restart ${remoteRoot}/www/current/ecosystem.config.js --env production --update-env && pm2 save'
+"
